requirements.txt
certifi==2020.6.20
chardet==3.0.4
Deprecated==1.2.10
idna==2.10
pkg-resources==0.0.0
PyGithub==1.53
PyJWT==1.7.1
requests==2.24.0
urllib3==1.25.10
wrapt==1.12.1
